--- a/src/test/resources/application.conf
+++ b/src/test/resources/application.conf
@@ -2,7 +2,7 @@ include "application.base.conf"
 
 system {
   services {
-    pocCreationInterval = 2
+    pocCreationInterval = 1
   }
 
   teamdrive {