--- a/src/test/scala/com/ubirch/e2e/E2ETestBase.scala
+++ b/src/test/scala/com/ubirch/e2e/E2ETestBase.scala
@@ -2,7 +2,7 @@ package com.ubirch.e2e
 
 import com.typesafe.scalalogging.StrictLogging
 import com.ubirch._
-import com.ubirch.db.tables.PocAdminRepository
+import com.ubirch.e2e.StartupHelperMethods.isPortInUse
 import com.ubirch.models.user.UserName
 import com.ubirch.services.jwt.PublicKeyPoolService
 import com.ubirch.services.keycloak.{
@@ -13,7 +13,6 @@ import com.ubirch.services.keycloak.{
 }
 import com.ubirch.services.poc.PocTestHelper.await
 import com.ubirch.services.{ CertifyKeycloak, DeviceKeycloak }
-import monix.eval.Task
 import monix.execution.Scheduler
 import org.flywaydb.core.Flyway
 import org.flywaydb.core.api.output.MigrateResult
@@ -47,8 +46,7 @@ trait E2ETestBase
         Random.alphanumeric.take(10).mkString(""))
     val injector = new E2EInjectorHelperImpl(superAdmin, tenantAdmin, useMockKeyDiscoveryService)
     KeyPoolServiceInitialization.initPool(injector)
-    InitDatabase.migration
-    cleanupDB()
+    prepareDB()
     testCode(injector)
   }
 
@@ -56,13 +54,23 @@ trait E2ETestBase
     Random.alphanumeric.take(10).mkString("")
   }
 
-  private def cleanupDB() = {
-    await(
-      StaticTestPostgresJdbcContext.ctx.executeAction(
-        "TRUNCATE poc_manager.poc_admin_table, poc_manager.poc_employee_table, poc_manager.poc_table, poc_manager.tenant_table, poc_manager.user_table, poc_manager.poc_admin_status_table, poc_manager.poc_employee_status_table, poc_manager.poc_status_table CASCADE"
-      ),
-      2.seconds
-    )
+  private def prepareDB() = {
+    if (isPortInUse(5432)) {
+      await(
+        StaticTestPostgresJdbcContext.ctx.executeAction(
+          "TRUNCATE poc_manager.poc_admin_table, poc_manager.poc_employee_table, poc_manager.poc_table, poc_manager.tenant_table, poc_manager.user_table, poc_manager.poc_admin_status_table, poc_manager.poc_employee_status_table, poc_manager.poc_status_table CASCADE"
+        ),
+        2.seconds
+      )
+    } else {
+      InitDatabase.migration
+      await(
+        StaticTestPostgresJdbcContext.ctx.executeAction(
+          "TRUNCATE poc_manager.poc_admin_table, poc_manager.poc_employee_table, poc_manager.poc_table, poc_manager.tenant_table, poc_manager.user_table, poc_manager.poc_admin_status_table, poc_manager.poc_employee_status_table, poc_manager.poc_status_table CASCADE"
+        ),
+        2.seconds
+      )
+    }
   }
 
   def performKeycloakCleanup(injector: E2EInjectorHelperImpl): Unit = {
@@ -110,19 +118,19 @@ object KeyPoolServiceInitialization {
 }
 
 object KeycloakCertifyContainer {
-  val container: KeycloakContainer =
+  lazy val container: KeycloakContainer =
     KeycloakContainer.Def(mountExtension = true, realmExportFile = "certify-export.json").start()
 }
 
 object KeycloakDeviceContainer {
-  val container: KeycloakContainer =
+  lazy val container: KeycloakContainer =
     KeycloakContainer.Def(mountExtension = false, realmExportFile = "device-export.json").start()
 }
 
 object PostgresDbContainer {
-  val container: PostgresContainer = PostgresContainer.Def().start()
+  lazy val container: PostgresContainer = PostgresContainer.Def().start()
 
-  val flyway: Flyway = Flyway
+  lazy val flyway: Flyway = Flyway
     .configure()
     .dataSource(
       s"jdbc:postgresql://${container.container.getContainerIpAddress}:${container.container.getFirstMappedPort}/postgres",
@@ -141,3 +149,27 @@ sealed trait DiscoveryServiceType {
 }
 case object MockDiscoveryService extends DiscoveryServiceType
 case object RealDiscoverService extends DiscoveryServiceType
+
+object StartupHelperMethods {
+  import java.io.IOException
+  import java.net.ServerSocket
+
+  def isPortInUse(port: Int): Boolean = {
+    var ss: ServerSocket = null
+    try {
+      ss = new ServerSocket(port)
+      ss.setReuseAddress(true)
+      return false
+    } catch {
+      case _: IOException =>
+
+    } finally {
+      if (ss != null)
+        try ss.close()
+        catch {
+          case _: IOException =>
+        }
+    }
+    true
+  }
+}