--- a/src/test/scala/com/ubirch/e2e/controllers/PocAdminControllerSpec.scala
+++ b/src/test/scala/com/ubirch/e2e/controllers/PocAdminControllerSpec.scala
@@ -627,9 +627,6 @@ class PocAdminControllerSpec
   override protected def beforeAll(): Unit = {
     super.beforeAll()
     withInjector { injector =>
-      lazy val pool = injector.get[PublicKeyPoolService]
-      await(pool.init(DeviceKeycloak, CertifyKeycloak), 2.seconds)
-
       lazy val superAdminController = injector.get[PocAdminController]
       addServlet(superAdminController, "/*")
     }