--- a/src/test/scala/com/ubirch/e2e/controllers/PocEmployeeControllerSpec.scala
+++ b/src/test/scala/com/ubirch/e2e/controllers/PocEmployeeControllerSpec.scala
@@ -209,9 +209,6 @@ class PocEmployeeControllerSpec extends E2ETestBase with BeforeAndAfterEach with
   override protected def beforeAll(): Unit = {
     super.beforeAll()
     withInjector { injector =>
-      lazy val pool = injector.get[PublicKeyPoolService]
-      await(pool.init(DeviceKeycloak, CertifyKeycloak), 2.seconds)
-
       lazy val pocEmployeeController = injector.get[PocEmployeeController]
       addServlet(pocEmployeeController, "/*")
     }