Following improvements were made:
* KeyPool is initialized only once
* Unnecessary calls to `withInjector` were removed. They were cleaning up and recreating the whole DB schema
* Instead of recreating the whole DB schema after each test, we are just truncating data from all tables
* Added functionality to wait until some condition is met instead of waiting for a fixed time
* Added functionality to wait for two ticks of *Loop services instead of waiting for a fixed time
* Added possibility to run the tests with services (Keycloaks, Postgres) running locally instead of creating them via testContainers. In order to run tests in such mode, just invoke `docker-compose up`. The functionality works via checking if ports (5432, 8080, 8081) are free. If they are free, then testContainers will start up new instances, if not, then we will try to connect to those instances directly. 
* Improved initializations of different services so it is done only once.

Numbers:
* Before the improvements, `mvn test` was taking around 4 minutes 30 seconds.
* After the improvements and in testContainers mode, `mvn test` takes around 2 minutes 20 seconds
* After the improvements and with local containers mode, `mvn test` takes around 1 minute 30 seconds.

* Before the improvements, running a single E2E test file was taking around 50 seconds (services startup time) + test time
* After the improvements and with local containers mode, running single E2E test does not require starting up services phase (-50 seconds to overall time). It enables to run them quickly while developing.